+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-04T12:19:55.618Z",
+    "2026-06-04T13:56:06.515Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {