+++ packages/orq-rc/RELEASES.md
@@ -9779,3 +9779,13 @@ Based on:
 - [typescript v4.10.0-rc.87] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.87] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.87 - packages/orq-rc
+
+## 2026-06-04 13:59:03
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.770.0 (2.893.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.88] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.88] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.88 - packages/orq-rc